// types/xmltv.i18n.ts
type XmltvStringWithLang = {
  /**
   * The value of the string
   */
  _value: string;

  /**
   * The lang code of the string eg. "en" for English, "es" for Spanish, etc.
   */
  lang?: string;
};

/**
 * The channel display name with an optional language attribute.
 */
export type XmltvDisplayName = XmltvStringWithLang;
// ...
export type XmltvIcon = {
  /**
   * URL of the icon
   */
  src: string;

  /**
   * Width of the icon
   */
  width?: number;

  /**
   * Height of the icon
   */
  height?: number;
};

/**
 * Object describing a URL, used in the programme and channel objects.
 */
export type XmltvUrl = {
  /**
   * Value of the URL
   */
  _value: string;

  /**
   * System of the URL, eg "imdb", "thetvdb", etc.
   */
  system?: string;
};
// ...
export type XmltvChannel = {
  /**
   * The channel id
   *
   * @example
   * ```typescript
   * {
   *   id: 'channel-1'
   * }
   * ```
   */
  id: string;

  /**
   * The channel display name
   *
   * @example
   * ```typescript
   * {
   *   displayName: [{
   *     _value: 'Channel 1',
   *     lang: 'en'
   *   },{
   *     _value: 'Kanal 1',
   *     lang: 'de'
   *   }]
   * }
   * ```
   */
  displayName: XmltvDisplayName[];

  /**
   * The channel icon
   *
   * @example
   * ```typescript
   * {
   *   icon: [{
   *     src: 'https://www.themoviedb.org/assets/1/v4/logos/32x32-blue-1f8b5c2fda197d0ee7d4f5b9fdca72a67ac3c9f7f8f8f8f8f8f8f8f8f8f8f8f8f.png',
   *     width: 32,
   *     height: 32
   *   }]
   * }
   * ```
   */
  icon?: XmltvIcon[];

  /**
   * The channel url
   *
   * @example
   * ```typescript
   * {
   *   url: [{_value: 'https://www.channel1.com'}]
   * ]
   * ```
   */
  url?: XmltvUrl[];
};

export type Xmltv = {
  channels?: XmltvChannel[];
  programmes?: XmltvProgramme[];
  date?: Date;
  sourceInfoName?: string;
  generatorInfoName?: string;
  sourceInfoUrl?: string;
  sourceDataUrl?: string;
  generatorInfoUrl?: string;
};